Moved tests into appropriate folder
